refactor: remove dead code from forked implementations
